fix tests and move import location
