+++ src/base-command.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 const { Command, flags } = require('@oclif/command')
-const Config = require('@adobe/aio-cli-config/lib/Config')
+const Config = require('@adobe/aio-cli-config/src/Config')
 const hjson = require('hjson')
 const yaml = require('js-yaml')
